handstand/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
mmengine>=0.10.0
mmcv>=2.0.0
mmdet>=3.3.0
mmpose>=1.3.0
ultralytics>=8.3.0
opencv-python
numpy>=1.23
scipy>=1.10
pandas>=2.0
tqdm>=4.66
rich>=13.5
yt-dlp>=2024.3.10
ffmpeg-python>=0.2.0
matplotlib>=3.8
pyyaml>=6.0
openai>=1.12.0
psutil>=5.9
imageio>=2.33
easydict>=1.10
ipdb>=0.13.13

